automation bug fix
fixed issue where inputting "Queen Elizabeth" or similar inputs would lead to an error, since there are multiple people with the same name. a proper error message is now displayed to the user. additionally, automation has been expanded to cases which previously did not work, such as with "basketball" and more.